+++ ontrack/src/components/BookPage/index..tsx
@@ -1,0 +1,1 @@
+export * from "./BookPage";
